moved rabbit to service to make the code structure clear.
