+++ src/modules/admin/message/message.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, OnModuleInit } from '@nestjs/common';
 import axios from 'axios';
-import { RabbitService } from './rabbit/rabbit.service';
+import { RabbitService } from './services/rabbit/rabbit.service';
 
 @Injectable()
 export class MessageService implements OnModuleInit{
