+++ src/modules/admin/message/services/rabbit/rabbit.module.ts
